1.6.3 typo bug fixed.

+++ tests/const/test1_derived.go
@@ -19,9 +19,9 @@ const (
  */
 // enum mColors
 const (
-    mColorsKBlueDef uint8 = 0
-    mColorsKGreenDef uint8 = 1
-    mColorsKRedDef uint8 = 2
+    mColorsKBlueDef uint16 = 0
+    mColorsKGreenDef uint16 = 1
+    mColorsKRedDef uint16 = 2
     mColorsKBlueStr string = "blue";
     mColorsKGreenStr string = "green";
     mColorsKRedStr string = "red";
